Use an unambiguous string to represent datetimes.
This sends a datetime to MySQL as 'YYYY-MM-DD hh:mm:ss' instead of the
numeric YYYYMMDDhhmmss.

This avoids problems when using small dates, such as year 1.  In that case
the previous code was sending 01 as the year part, which MySQL parses as
2001, not 1.

The accepted formats are documented at
http://dev.mysql.com/doc/refman/5.7/en/date-and-time-literals.html.
Maybe one day MySQL will support ISO 8601 *sigh*